web-fire-app/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
Pillow==10.0.1
folium==0.14.0
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1
opencv-python==4.8.1.78
geopy==2.4.0
rasterio==1.3.9
geopandas==0.14.1
xarray==2023.9.0
netCDF4==1.6.4
matplotlib==3.7.2
scikit-image==0.21.0
gunicorn==21.2.0
aiohttp==3.9.1
